+++ apps/website/scripts/deprecations/deprecations.ts
@@ -1,5 +1,6 @@
 import { Q22023 } from './configs/Q22023';
 import { Q32023 } from './configs/Q32023';
+import { Q42023 } from './configs/Q42023';
 import { Deprecation } from './types';
 
-export const deprecations: Deprecation[] = [Q22023, Q32023];
+export const deprecations: Deprecation[] = [Q22023, Q32023, Q42023];