+++ packages/cds-migrator/src/generators/migrate/migrate.ts
@@ -1,5 +1,6 @@
 import { Tree } from '@nrwl/devkit';
 
+import migrateToQ42023Deprecations from '../../migrations/q42023';
 import migrateTo4_0_0 from '../../migrations/update-4-0-0';
 import migrateTo5_0_0 from '../../migrations/update-5-0-0';
 import migrateToDecompedPackages from '../../migrations/update-decomp';
@@ -22,6 +23,10 @@ async function migrate(tree: Tree, options: MigrateOptions) {
 
     case 'package-decomp':
       await migrateToDecompedPackages(tree);
+      break;
+
+    case 'q42023-deprecations':
+      await migrateToQ42023Deprecations(tree);
       break;
 
     default: