+++ packages/cds-migrator/src/migrations/q42023/index.ts
@@ -1,0 +1,7 @@
+import { Tree } from '@nrwl/devkit';
+
+import componentMigrations from './componentMigrations';
+
+export default async function main(tree: Tree) {
+  await componentMigrations(tree);
+}